+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Administration/QueueProperties.cs
@@ -147,7 +147,7 @@ namespace Azure.Messaging.ServiceBus.Administration
         /// <summary>
         /// Indicates whether this queue has dead letter support when a message expires.
         /// </summary>
-        /// <remarks>If true, the expired messages are moved to dead-letter sub-queue. Default value is false.</remarks>
+        /// <remarks>If true, the expired messages are moved to dead-letter subqueue. Default value is false.</remarks>
         public bool DeadLetteringOnMessageExpiration { get; set; } = false;
 
         /// <summary>
@@ -268,7 +268,7 @@ namespace Azure.Messaging.ServiceBus.Administration
         public bool EnablePartitioning { get; internal set; }
 
         /// <summary>
-        /// Custom metdata that user can associate with the description.
+        /// Custom metadata that user can associate with the description.
         /// </summary>
         /// <remarks>Cannot be null. Max length is 1024 chars.</remarks>
         public string UserMetadata