+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Common/src/Shared/Queues/SharedQueueWatcher.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Concurrent;
 
-namespace Microsoft.Azure.WebJobs.Host.Queues.Listeners
+namespace Microsoft.Azure.WebJobs.Extensions.Storage.Common.Listeners
 {
     internal class SharedQueueWatcher : IMessageEnqueuedWatcher
     {