+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Queues/samples/QueueExtensionSamples.cs
@@ -5,15 +5,14 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Azure.WebJobs.Extensions.Storage.Common.Tests;
-using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Storage.Common.Tests;
 using Microsoft.Extensions.Azure;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
 
-namespace Azure.WebJobs.Extensions.Storage.Blobs.Samples.Tests
+namespace Microsoft.Azure.WebJobs.Extensions.Storage.Queues.Samples.Tests
 {
     public class BlobExtensionSamples
     {